+++ packages/react/runtime/__test__/utils/envManager.ts
@@ -11,7 +11,7 @@ import { globalPipelineOptions, setPipeline } from '../../src/lynx/performance.j
 import { __root, setRoot } from '../../src/root.js';
 import { SnapshotInstance, backgroundSnapshotInstanceManager, snapshotInstanceManager } from '../../src/snapshot.js';
 import { hydrationMap } from '../../src/snapshotInstanceHydrationMap.js';
-import { clearWorkletRefLastIdForTesting } from '../../src/worklet/workletRef.js';
+import { clearWorkletRefLastIdForTesting } from '../../src/worklet/ref/workletRef.js';
 
 export class EnvManager {
   root: typeof __root | undefined;
