+++ packages/react/runtime/src/worklet/destroy.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-import { takeWorkletRefInitValuePatch } from './workletRefPool.js';
+import { takeWorkletRefInitValuePatch } from './ref/workletRefPool.js';
 
 export const destroyTasks: (() => void)[] = [];
