+++ packages/react/runtime/src/worklet/ref/workletRefPool.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-import { isMtsEnabled } from './functionality.js';
+import { isMtsEnabled } from '../functionality.js';
 
 let initValuePatch: [number, unknown][] = [];
 const initValueIdSet = /*#__PURE__*/ new Set<number>();
